pyproject.toml
[tool.poetry]
name = "settingsd"
version = "0.1.0"
description = ""
authors = ["Nicholas Kostelnik <nitaigao@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dbus-python = "^1.2"
PyGObject = "^3.34"

[tool.poetry.dev-dependencies]
pylint = "^2.4"
pytest = "^5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
